+++ lib/ui/screen/MainScreen.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_app/generated/R.dart';
-import 'package:flutter_app/src/costants/app_colors.dart';
 import 'package:flutter_app/src/file/MyDownloader.dart';
 import 'package:flutter_app/src/util/LanguageUtil.dart';
 import 'package:flutter_app/src/store/Model.dart';
@@ -9,8 +8,8 @@ import 'package:flutter_app/src/taskcontrol/task/CheckCookiesTask.dart';
 import 'package:flutter_app/ui/other/MyToast.dart';
 import 'package:flutter_app/ui/pages/coursetable/CourseTablePage.dart';
 import 'package:flutter_app/ui/pages/mail/NewAnnouncementPage.dart';
+import 'package:flutter_app/ui/pages/other/OtherPage.dart';
 import 'package:flutter_app/ui/pages/score/ScorePage.dart';
-import 'package:flutter_app/ui/pages/setting/SettingPage.dart';
 import 'package:eva_icons_flutter/eva_icons_flutter.dart';
 
 
@@ -38,7 +37,7 @@ class _MainScreenState extends State<MainScreen> {
       _pageList.add(NewAnnouncementPage());
       //bottomPageList.add(BottomPage(CalendarScreen()));
       _pageList.add(ScoreViewerPage());
-      _pageList.add(SettingPage(_pageController));
+      _pageList.add(OtherPage(_pageController));
       _setLang();
     });
     _flutterDownloaderInit();
